// src/config/constants/types.ts
export interface Address {
  256256?: string
  512512: string
}

export interface SerializedToken {
  chainId: number
  address: string
  decimals: number
  symbol?: string
  name?: string
  projectLink?: string
}
// ...
export enum PrivateSaleType {
  seedsale = 'seedSale',
  privatesale = 'privateSale',
  preSale = 'presSale',
}

export interface PrivateSaleBaseProps {
  type: PrivateSaleType
  manager: Address,
  price: number
  name: string
  desc: string
  whitelistEnabled: boolean
}
export interface SerializedPrivateSaleConfig extends PrivateSaleBaseProps {
  tempToken: SerializedToken
  quoteToken: SerializedToken
  startDate?: number
  endDate?: number
  claimStartDate?: number
  claimEndDate?: number
  claimDays?: number[]
  claimPercents?: number[]
}